Ep40 User Profile Pagination Part 1
Pagination often comes up when working with Firebase and building social media applications.  The reason for this is because pagination allows us to more quickly fetch data from Firebase Database.  In this lesson, I'll walk you through how to implement this very tricky feature.  First we'll tackle how to paginate through Post objects in the User Profile screen.  In order to do so, we implement a few tricks on cellForItemAt to trigger the pagination process.

Once you complete this lesson, you'll be ready to move onto paginating through data and sorting based on a key such as creationDate on our Post model.