Setting up Push Notifications with FCM Instagram Firebase
Push Notifications are an essential part of increasing engagement of users for your applications.  With the help of Firebase Cloud Messaging, it is very easy to register your devices and send segmented notifications or individual notifications to your users.

In today's lesson, we will be registering an App ID in Apple's Developer console to allow our application to receive push notifications.  This process requires a developer's account with Apple that costs $99/yr.  There isn't any way around this and it's money developers have to dish out every year.